requirements.txt
fastapi
uvicorn
chromadb
python-dotenv
google-genai
streamlit